requirements.txt
# AURA — requirements.txt (pinned for stability)

#  Web Framework 
Flask>=2.3,<4
python-dotenv>=1.0

#  Security & Rate Limiting 
flask-talisman>=1.1
flask-limiter>=3.5

#  Networking & API Requests 
requests>=2.31
urllib3>=2.0
certifi
idna
charset-normalizer

#  Audio Input/Output (Voice) 
SpeechRecognition>=3.10
pyaudio
pyttsx3>=2.90
sounddevice
playsound3

#  Windows-Specific Audio & Controls 
pywin32
pypiwin32
comtypes

#  AI / LLM 
groq>=0.9
pyngrok

#  Database 
pymongo[srv]>=4.6       # MongoDB Atlas (SRV connection string support)

#  Vision / Neural Networks 
torch>=2.0
torchvision
timm
ultralytics>=8.0

#  Images & OCR 
opencv-python
opencv-contrib-python
pytesseract
Pillow>=10.0

#  Safe Expression Evaluation 
simpleeval>=0.9.13

#  Automation & System 
pywhatkit
pyautogui
psutil>=5.9
numpy>=1.24
wikipedia